requirements.txt
fastapi
uvicorn
tensorflow
joblib
pandas
numpy
scikit-learn
Pillow
